+++ packages/reakit/src/Arrow/Arrow.ts
@@ -1,7 +1,5 @@
-import { ComponentType } from "react";
-import PropTypes from "prop-types";
+import * as PropTypes from "prop-types";
 import { prop } from "styled-tools";
-import globalPropTypes from "../_utils/globalPropTypes";
 import styled from "../styled";
 import as from "../as";
 import Base from "../Base";
@@ -9,10 +7,10 @@ import Base from "../Base";
 const SQRT2 = 1.41421356237;
 
 interface ArrowProps {
-  angle?: number | null;
+  angle?: number;
 }
 
-const Arrow = styled(Base as ComponentType<ArrowProps>)`
+const Arrow = styled(Base)<ArrowProps>`
   display: inline-flex;
   width: 1em;
   height: 1em;
@@ -40,8 +38,8 @@ const Arrow = styled(Base as ComponentType<ArrowProps>)`
   ${prop("theme.Arrow")};
 `;
 
+// @ts-ignore
 Arrow.propTypes = {
-  ...globalPropTypes,
   angle: PropTypes.number
 };