+++ packages/reakit/src/Base/Base.ts
@@ -1,33 +1,30 @@
-import React, { ComponentType } from "react";
-import PropTypes from "prop-types";
+import * as React from "react";
+import * as PropTypes from "prop-types";
 import { prop } from "styled-tools";
 import { bool } from "../_utils/styledProps";
-import globalPropTypes from "../_utils/globalPropTypes";
 import styled from "../styled";
 import as from "../as";
 
-enum Position {
-  static = "static",
-  absolute = "absolute",
-  fixed = "fixed",
-  relative = "relative",
-  sticky = "sticky"
-}
-
-type PositionProps = { [key in keyof typeof Position]?: boolean | null };
-
-type ComponentProps = {
-  as: keyof JSX.IntrinsicElements | ComponentType;
+const positions = {
+  static: "static",
+  absolute: "absolute",
+  fixed: "fixed",
+  relative: "relative",
+  sticky: "sticky"
 };
 
-const positions = Object.keys(Position);
+type ComponentProps = {
+  as: keyof JSX.IntrinsicElements | React.ComponentType;
+};
 
 const Component = ({ as: T, ...props }: ComponentProps) =>
   React.createElement(T, props);
 
-type BaseProps = PositionProps & ComponentProps;
+type BaseProps = { [key in keyof typeof positions]?: boolean };
 
-const Base = styled<BaseProps>(Component)`
+const positionsKeys = Object.keys(positions);
+
+const Base = styled(Component)<BaseProps>`
   margin: 0;
   padding: 0;
   border: 0;
@@ -37,7 +34,7 @@ const Base = styled<BaseProps>(Component)`
   box-sizing: border-box;
   ${prop("theme.Base")};
   &&& {
-    ${bool("position", positions)};
+    ${bool("position", positionsKeys)};
   }
 `;
 
@@ -45,7 +42,6 @@ const asTypes = [PropTypes.func, PropTypes.string];
 
 // @ts-ignore
 Base.propTypes = {
-  ...globalPropTypes,
   as: PropTypes.oneOfType([
     ...asTypes,
     PropTypes.arrayOf(PropTypes.oneOfType(asTypes))