+++ packages/reakit/src/as.tsx
@@ -1,38 +1,38 @@
 /* eslint-disable no-param-reassign */
-import React, { ReactElement, ComponentType, HTMLProps, Ref } from "react";
+import * as React from "react";
 import pickCSSProps from "./_utils/pickCSSProps";
 import parseTag from "./_utils/parseTag";
 import parseClassName from "./_utils/parseClassName";
 import pickHTMLProps from "./_utils/pickHTMLProps";
 import CSSProps from "./_utils/CSSProps";
+import getComponentName from "./_utils/getComponentName";
 import { isStyledComponent } from "./styled";
 
-export type Omit<T, K extends keyof T> = Pick<
-  T,
-  ({ [P in keyof T]: P } & { [P in K]: never })[keyof T]
->;
+type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
 export type CSSProperties = {
   [key in keyof typeof CSSProps]?: string | number
 };
 
-export type AllProps<T = any> = Omit<HTMLProps<T>, "as"> &
+export type AllProps<T = any> = Omit<React.HTMLProps<T>, "as"> &
   ReaKitProps<T> &
   CSSProperties;
 
 export type SingleAsProp =
   | keyof JSX.IntrinsicElements
-  | ComponentType<AllProps>;
+  | React.ComponentType<AllProps>;
 
 export type AsProp = SingleAsProp | SingleAsProp[];
 
 export interface ReaKitProps<T = any> {
   as?: AsProp | null;
   nextAs?: SingleAsProp | null;
-  elementRef?: Ref<T>;
+  elementRef?: React.Ref<T>;
 }
 
-export type ReaKitComponent<P extends AllProps = {}> = ComponentType<P> & {
+export type ReaKitComponent<P extends AllProps = object> = React.ComponentType<
+  P
+> & {
   asComponents: AsProp;
   as: (asComponents: AsProp) => ReaKitComponent<P>;
 };
@@ -41,7 +41,7 @@ function As({ nextAs, ...props }: AllProps) {
   return render({ ...props, as: nextAs });
 }
 
-function render({ as: t, ...props }: AllProps): ReactElement<any> | null {
+function render({ as: t, ...props }: AllProps): JSX.Element {
   const T = parseTag(t);
 
   if (Array.isArray(T)) {
@@ -82,18 +82,16 @@ function isWrappedWithAs(target: any): target is ReaKitComponent {
   return typeof target.asComponents !== "undefined";
 }
 
-type AllPropsWithAs<T extends AllProps> = Omit<T, "as"> & AllProps;
+type AllPropsReplaceAs<P extends AllProps> = Omit<P, "as"> & AllProps;
 
 function as(asComponents: AsProp) {
-  return <P extends AllProps>(
-    WrappedComponent: ComponentType<P>
-  ): ReaKitComponent<AllPropsWithAs<P>> => {
+  return <P extends AllProps>(WrappedComponent: React.ComponentType<P>) => {
     const target = isStyledComponent(WrappedComponent)
       ? WrappedComponent.target
       : WrappedComponent;
 
     const defineProperties = (scope: typeof WrappedComponent) => {
-      const xscope = scope as ReaKitComponent<AllPropsWithAs<P>>;
+      const xscope = scope as ReaKitComponent<AllPropsReplaceAs<P>>;
       xscope.asComponents = asComponents;
       xscope.as = otherComponents => as(otherComponents)(scope);
       return xscope;
@@ -103,17 +101,19 @@ function as(asComponents: AsProp) {
       return defineProperties(WrappedComponent);
     }
 
-    const getComponentName = (component: typeof WrappedComponent): any =>
-      component.displayName || component.name || component;
-
-    const components = ([] as any[]).concat(WrappedComponent, asComponents);
-
     const getAs = (props: AllProps): SingleAsProp[] =>
-      components.concat(props.as || [], props.nextAs || []);
+      ([] as any[]).concat(
+        WrappedComponent,
+        asComponents,
+        props.as || [],
+        props.nextAs || []
+      );
 
-    const displayName = `${getComponentName(
-      WrappedComponent
-    )}.as(${([] as any[]).concat(asComponents).map(getComponentName)})`;
+    const componentName = getComponentName(WrappedComponent);
+    const commaSeparatedAs = ([] as any[])
+      .concat(asComponents)
+      .map(getComponentName);
+    const displayName = `${componentName}.as(${commaSeparatedAs})`;
 
     const EnhancedComponent: typeof WrappedComponent = (props: AllProps) =>
       render({ ...props, as: getAs(props) });