+++ types/index.d.ts
@@ -1,3 +1,1 @@
-declare module "styled-tools";
-
-declare module "constate";
+declare const dataLayer: object[];