+++ CHANGELOG.md
@@ -3,8 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.14.4"></a>
+## [0.14.4](https://github.com/reakit/reakit/compare/reakit@0.14.3...reakit@0.14.4) (2018-08-23)
+
+
+### Bug Fixes
+
+* TypeScript definition errors ([#201](https://github.com/reakit/reakit/issues/201)) ([f16de0f](https://github.com/reakit/reakit/commit/f16de0f)), closes [#198](https://github.com/reakit/reakit/issues/198)
+
+
+### Performance Improvements
+
+* Replace `validAttr` by `@emotion/is-prop-valid` ([#200](https://github.com/reakit/reakit/issues/200)) ([4a708a0](https://github.com/reakit/reakit/commit/4a708a0)), closes [#173](https://github.com/reakit/reakit/issues/173)
+
+
+
+
+
+<a name="0.14.3"></a>
+## [0.14.3](https://github.com/reakit/reakit/compare/reakit@0.14.2...reakit@0.14.3) (2018-08-19)
+
+
+### Features
+
+* Add `shift` and `flip` props to `Popover` ([#196](https://github.com/reakit/reakit/issues/196)) ([9ee52ed](https://github.com/reakit/reakit/commit/9ee52ed))
+* Add more TypeScript typings ([#193](https://github.com/reakit/reakit/issues/193)) ([b6c08fe](https://github.com/reakit/reakit/commit/b6c08fe))
+
+
+
+
+
+<a name="0.14.2"></a>
+## [0.14.2](https://github.com/reakit/reakit/compare/reakit@0.14.1...reakit@0.14.2) (2018-08-16)
+
+
+### Bug Fixes
+
+* `Arrow` and JS files typings not working ([#192](https://github.com/reakit/reakit/issues/192)) ([ce88103](https://github.com/reakit/reakit/commit/ce88103))
+
+
+
+
+
 <a name="0.14.1"></a>
-## 0.14.1 (2018-08-15)
+## [0.14.1](https://github.com/reakit/reakit/compare/v0.14.0...reakit@0.14.1) (2018-08-15)
 
 
 ### Features
@@ -17,7 +59,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 <a name="0.14.0"></a>
-# 0.14.0 (2018-08-09)
+# [0.14.0](https://github.com/reakit/reakit/compare/v0.13.2...v0.14.0) (2018-08-09)
 
 
 ### Features
@@ -71,7 +113,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 <a name="0.13.2"></a>
-## 0.13.2 (2018-08-08)
+## [0.13.2](https://github.com/reakit/reakit/compare/v0.13.1...v0.13.2) (2018-08-08)
 
 
 ### Bug Fixes
@@ -88,7 +130,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 <a name="0.13.1"></a>
-## 0.13.1 (2018-08-05)
+## [0.13.1](https://github.com/reakit/reakit/compare/v0.13.0...v0.13.1) (2018-08-05)
 
 
 ### Bug Fixes
@@ -103,7 +145,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 <a name="0.13.0"></a>
-# 0.13.0 (2018-08-03)
+# [0.13.0](https://github.com/reakit/reakit/compare/v0.12.1...v0.13.0) (2018-08-03)
 
 
 ### Features