+++ packages/website/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.14.4"></a>
+## [0.14.4](https://github.com/reakit/reakit/compare/website@0.14.3...website@0.14.4) (2018-08-23)
+
+
+### Bug Fixes
+
+* TypeScript definition errors ([#201](https://github.com/reakit/reakit/issues/201)) ([f16de0f](https://github.com/reakit/reakit/commit/f16de0f)), closes [#198](https://github.com/reakit/reakit/issues/198)
+
+
+
+
+
+<a name="0.14.3"></a>
+## [0.14.3](https://github.com/reakit/reakit/compare/website@0.14.2...website@0.14.3) (2018-08-19)
+
+
+### Features
+
+* **website:** Add analytics trackings ([#195](https://github.com/reakit/reakit/issues/195)) ([cc697c6](https://github.com/reakit/reakit/commit/cc697c6))
+* **website:** Add newsletter form ([#194](https://github.com/reakit/reakit/issues/194)) ([3e568ff](https://github.com/reakit/reakit/commit/3e568ff))
+* Add `shift` and `flip` props to `Popover` ([#196](https://github.com/reakit/reakit/issues/196)) ([9ee52ed](https://github.com/reakit/reakit/commit/9ee52ed))
+
+
+
+
+
+<a name="0.14.2"></a>
+## [0.14.2](https://github.com/reakit/reakit/compare/website@0.14.1...website@0.14.2) (2018-08-16)
+
+**Note:** Version bump only for package website
+
+
+
+
+
 <a name="0.14.1"></a>
 ## 0.14.1 (2018-08-15)